==> code/01_data-cleaning.ipynb <==
# Cleaning the Tweets for EDA

## Imports

import pandas as pd
from collections import Counter

# READ CSV 
df = pd.read_csv("../datasets/raw_df.csv")

## Clean Data

### Remove Unnecessary Columns

Only the `username` and `text` columns will be kept.

df = df[["username", "text"]]

df.head(5)

### Clean HTML,  Make Lowercase, Remove Links

Many of the texts contain non-letters, uppercase letters, `URL` links, or `hashtags` that will need to be removed for EDA.

**Example Below:**

df["text"][4]

def clean_df(data, col):
    
    # REMOVE URL LINKS 
    data[col] = data[col].replace(r'http\S+', '', regex=True).replace(r'www\S+', '', regex=True) 
    # https://stackoverflow.com/questions/51994254/removing-url-from-a-column-in-pandas-dataframe
    
    # REMOVE NON-LETTERS:
    data[col] = data[col].str.replace('[^a-zA-Z]', ' ')
    
    # MAKE LOWERCASE:
    data[col] = data[col].str.lower()
    
    return data

clean_df(df, "text")
clean_df(df, "username")

df["text"][4]

df.to_csv("../datasets/df.csv")

# NEEDS FURTHER IMPLEMENTATION
# df.to_sql("../datasets/twitter_text", con=engine)

## Data for EDA

### Create Keywords

# USING PREBUILT CRISIS LEXICON KEYWORDS
    # https://crisislex.org/crisis-lexicon.html
def create_keywords_list(filepath):
    keywords = open(filepath)
    clean_text = ""
    clean_keywords = []
    keywords_list = []
    result = []
    
    for keyword in keywords:
        keywords_list.append(keyword)

    for words in keywords_list:
        clean_text = words.replace("\n", "")
        clean_keywords.append(clean_text.split())

    for i in clean_keywords:
        for j in i:
            result.append(j)
    return result 

keywords = create_keywords_list("../datasets/CrisisLexLexicon/CrisisLexRec.txt")

def check_keywords(data, col, lst):
    kw_df = pd.DataFrame()
    
    for word in data[col]:
        if word in lst:
            kw_df = kw_df.append(data.loc[data[col].str.contains(word)], ignore_index=True)
    return kw_df

### Explore Usernames

# MAJORITY USERNAMES ARE EMS SERVICES
# THERE ARE BOTS
# THERE ARE EMOJIS

# CREATE A TABLE OF 100 MOST COMMON WORDS IN USERNAMES
Counter(" ".join(df["username"]).split()).most_common(100)
username_most_occur_df = pd.DataFrame(Counter(" ".join(df["username"]).split()).most_common(100),
                          columns=["username", "no_occurences"])

username_most_occur_df

username_most_occur_df.to_csv("../datasets/username-most-freq.csv")

username_keywords = check_keywords(username_most_occur_df, "username", keywords)
username_keywords

username_keywords.to_csv("../datasets/username-crisisword-match.csv")

### Explore Text

# CREATE A TABLE OF 100 MOST COMMON WORDS IN THE TEXT OF EACH TWEET
text_occur_df = pd.DataFrame(Counter(" ".join(df["text"]).split()).most_common(100),
                             columns=["text", "no_occurences"])
text_occur_df

text_occur_df.to_csv("../datasets/text-most-freq.csv")

text_keywords = check_keywords(text_occur_df, "text", keywords)
text_keywords

text_keywords.to_csv("../datasets/text-crisisword-match.csv")
