==> code/03_bagofwords.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Imports" data-toc-modified-id="Imports-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Imports</a></span></li><li><span><a href="#Importing-dataset-of-labeled-diaster-messages" data-toc-modified-id="Importing-dataset-of-labeled-diaster-messages-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Importing dataset of labeled diaster messages</a></span></li><li><span><a href="#Processing-Messages" data-toc-modified-id="Processing-Messages-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Processing Messages</a></span></li><li><span><a href="#Function-for-modeling" data-toc-modified-id="Function-for-modeling-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Function for modeling</a></span></li><li><span><a href="#Modeling-Tests" data-toc-modified-id="Modeling-Tests-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Modeling Tests</a></span></li><li><span><a href="#Final-Model" data-toc-modified-id="Final-Model-6"><span class="toc-item-num">6&nbsp;&nbsp;</span>Final Model</a></span></li><li><span><a href="#Coefficients" data-toc-modified-id="Coefficients-7"><span class="toc-item-num">7&nbsp;&nbsp;</span>Coefficients</a></span></li></ul></div>

# Imports 

import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt

import nltk
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer
import regex as re
import unidecode
from nltk.corpus import stopwords

#For Modeling
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix
from sklearn.svm import SVC
from sklearn.metrics import roc_auc_score
from sklearn.preprocessing import StandardScaler


# Importing dataset of labeled diaster messages 

#read in the open source disaster response messages dataset 
df = pd.read_csv('..//datasets/disaster_response_messages_training.csv', header = 0)

df.head()

# Processing Messages

stops = stopwords.words('english')

def message_to_words(raw_message):
    
     # remove accents
    unaccented = unidecode.unidecode(raw_message)
    
    # remove all non-letter characters
    letters_only = re.sub("[^a-zA-Z]", " ", unaccented)
    
    # lowercase 
    words = letters_only.lower().split()
    
    # lemmatize
    lemmatizer = WordNetLemmatizer()
    tokens_lem = [lemmatizer.lemmatize(i) for i in words]
    
    # stop words
    meaningful_words = [w for w in words if not w in stops]
    
    # return as a string 
    return(" ".join(meaningful_words))

total_message = df.shape[0]
clean_message = []

print("Cleaning and parsing the message...")

j = 0
for message in df['message']:
    clean_message.append(message_to_words(message))
    
    # If the index is divisible by 100, print a message
    if (j+1) % 100 == 0:
        print(f'Comment {j+1} of {total_message}.')
    
    j += 1
    
    if j == total_message:
        print('Done.')

df = df.assign(cleaned_message = clean_message)
df.head(3)

# Function for modeling 

##Taken from NYC I Example - Building off of their work to make our words2vec model stronger (Dan Kim, Cameron 
#Bronstein, Patrick O'Connell)

def text_to_model(X_column, model, vectorizer, params, verbose = 1):
    
    X = df[X_column]                                    #creates X and y
    y = df['direct_report']
    
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42, test_size = .25)    #train test split
    
    pipe = Pipeline([                               #pipeline to run with gridsearch and test hyperparameters
        ('vectorizer', vectorizer),                 #for both the model and vectorizer.
        ('model', model)])                          #this will be done for many vectorizer-model combinations
        
    grid = GridSearchCV(pipe, param_grid=params, cv=5, verbose=verbose)  
    
    grid.fit(X_train, y_train)                       #fitting the grid model to X_train and y_train and 
                                                     #running a 5 fold cross validation
    score_dict = {}
    
    score_dict['X'] = X_column                               #this dict will be converted to dataframe to store 
    score_dict['Vectorizer'] = vectorizer                    #the performance of each gridsearch and return the 
    score_dict['Model'] = model                              #best parameters and score to compare to other models
    score_dict['train_score'] = grid.score(X_train, y_train)
    score_dict['test_score'] = grid.score(X_test, y_test)
    score_dict['best_params'] = grid.best_params_
    
    try:
        return pd.DataFrame(score_dict)
    except:
        return score_dict

# Modeling Tests

# Text Selected - Tokenized messsages
# Running a gridsearch on Logistic Regression and Count Vectorizer 

logreg = LogisticRegression()
cv = CountVectorizer()

params = {
          'vectorizer__max_features':[100, 500, 1000, 1500, 5000],
          'vectorizer__ngram_range':[(1,1), (1,2), (1,3)],
          'vectorizer__min_df':[1, 2, 3, 4],
          'vectorizer__max_df':[1.0],
          'model__penalty':['l2'],
          'model__max_iter': [1500]
         }

count_vect_logreg = text_to_model('cleaned_message', model = logreg, vectorizer=cv, params=params)
count_vect_logreg

# Text Selected - Tokenized Message
# Gridsearching Hyperparameters for TF-IDF Vectorizer

logreg = LogisticRegression()
tfidf = TfidfVectorizer()

params = {
          'vectorizer__max_features':[1000, 1500, 5000, 50000],
          'vectorizer__ngram_range':[(1,1), (1,2), (1,3)],
          'vectorizer__min_df':[1, 2, 3, 4],
          'vectorizer__max_df':[1.0],
          'model__penalty':['l2'],
          'model__max_iter': [1500]
         }

tfidf_vect_logreg = text_to_model(X_column='cleaned_message', model = logreg, vectorizer=tfidf, params=params)
tfidf_vect_logreg

# Final Model

X = df['cleaned_message']                                   
y = df['direct_report']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42, test_size = .2) 

#After grid searching I choose to fit and create predictions with my best fit Logistic Regression model
vec = TfidfVectorizer({ 'max_features':[50000],
                       'ngram_range':(1,2),
                       'min_df':[3],
                       'max_df':[1.0]})

# Scale data with vectorizer
X_train_vec = vec.fit_transform(X_train)
X_train_vec = pd.DataFrame(X_train_vec.toarray(), columns=vec.get_feature_names())

# Transform the test set
X_test_vec = vec.transform(X_test)

# Instantiate model
model = LogisticRegression(solver='lbfgs', penalty = 'l2', max_iter = 1500)

# Fit on training data.
model.fit(X_train_vec, y_train)

# Get scores
print('CV score:', cross_val_score(model, X_train_vec, y_train, cv=5).mean())
print('Training Score:', model.score(X_train_vec, y_train))
print('Testing Score:', model.score(X_test_vec, y_test))

pred = model.predict(X_test_vec)

cm = confusion_matrix(y_test, pred)

cm_df = pd.DataFrame(cm, columns=['predneg', 'pred pos'], index=['actual neg', 'actual pos'])
cm_df

tn, fp, fn, tp = confusion_matrix(y_test, pred).ravel()

accuracy = (tp + tn) / (tp + fn + fp + tn) * 100
misclassification = (100 - accuracy)
sensitivity = tp / (tp + fn) * 100
specificity = tn / (tn + fp) * 100
precision = tp / (tp + fp) * 100
roc_auc = roc_auc_score(y_test, pred)

print(f'Accuracy: {round(accuracy, 4)}%')
print(f'Misclassification rate: {round(misclassification, 4)}%')
print(f'Sensitivity: {round(sensitivity, 4)}%')
print(f'Specificity: {round(specificity, 4)}%')
print(f'Precision: {round(precision, 4)}%')
print(f'ROC AUC: {round(roc_auc, 4)}')

# Coefficients 

#Create dataframe with coefs and e^coefs for each word

coefs = list(zip(vec.get_feature_names(), model.coef_[0].T))
coefs = pd.DataFrame(coefs, columns = ['word','coef'])
coefs['e^coef'] = np.exp(coefs['coef'])

coefs.sort_values(by='e^coef', ascending=False).head(30)

coefs.sort_values(by='e^coef').head(20)
