==> code/04_word2vec.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Imports" data-toc-modified-id="Imports-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Imports</a></span></li><li><span><a href="#Bag-of-Words" data-toc-modified-id="Bag-of-Words-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Bag of Words</a></span></li><li><span><a href="#Train-Words2Vec" data-toc-modified-id="Train-Words2Vec-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Train Words2Vec</a></span></li><li><span><a href="#Importing-the-tweets" data-toc-modified-id="Importing-the-tweets-4"><span class="toc-item-num">4&nbsp;&nbsp;</span>Importing the tweets</a></span></li><li><span><a href="#Creating-the-Urgent-vs.-non-Urgent-Vectors" data-toc-modified-id="Creating-the-Urgent-vs.-non-Urgent-Vectors-5"><span class="toc-item-num">5&nbsp;&nbsp;</span>Creating the Urgent vs. non-Urgent Vectors</a></span></li></ul></div>

# Imports 

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt 

import nltk
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer
import regex as re
import unidecode
from nltk.corpus import stopwords

import gensim

# Bag of Words

- **words were created using a logistic regression on an open source database used for NLP that contains messages during disasters and their categorizations, one of those being if the message was a direct call for help, what we will classify as a message being more "urgent"**
    - https://appen.com/datasets/combined-disaster-response-data/
    
    
- **another way to classify is a message is more urgent is using FEMA's Public Assistance Program and Policy Guide**
    - in this guide, public assistance work is either emergency work or permanent work. Important terms from each designation were added to our bag of words 
    - **Emergency**: 
        - Emergency Protective Measures
        - Debris Removal 
    - **Permanent**:
        - Roads and Bridges
        - Water Control Facilities
        - Buildings and Equipment
        - Utilities 
        - Parks, Recreational, other 
        
    - https://www.fema.gov/media-library-data/1525468328389-4a038bbef9081cd7dfe7538e7751aa9c/PAPPG_3.1_508_FINAL_5-4-2018.pdf
   

#list of urgent words
urgent = ['help',
          'campfire',
          'tonight',
          'today',
          'fire',
          'need',
          'aid',
          'removal',
          'burn',
          'tree',
          'evacuation',
          'smoke',
          'unsafe',
          'search',
          'lost',
          'victim',
          'medical',
          'med',
          'urgent',
          'home',
          'haze',
          'serious',
          'fires',
          'reporting',
          'smoking',
          'come',
          'emergency',
          'wildfire',
          'Portland',
          'send',
          'park',
          'valley',
          'investigation',
          'structure'
         ]

#list of non-urgent words 
non_urgent = ['job',
              'news',
              'government',
              'country',
              'materials',
              'work',
              'price',
              'utilities',
              'facility',
              'erosion',
              'irrigation',
              'restoration',
              'shoulder',
              'mud',
              'silt',
              'ditch',
              'slip',
              'pray',
              'hope',
              'thanks',
              'thankful'
             ]

# Train Words2Vec

#Model is commented out because the google news dataset is too large to fit on git
#model = gensim.models.KeyedVectors.load_word2vec_format('..//datasets/GoogleNews-vectors-negative300.bin', binary=True)

#Checking vector size of GoogleNews Word2Vec
model.vector_size

# Importing the tweets 

df = pd.read_csv('..//datasets/tweets_lemmatized.csv')

df.head()

stops = stopwords.words('english')

#Cleaning messages to ensure that they are adequately prepped for analysis 
def message_to_words(raw_message):
    
     # remove accents
    unaccented = unidecode.unidecode(raw_message)
    
    # remove all non-letter characters
    letters_only = re.sub("[^a-zA-Z]", " ", unaccented)
    
    # lowercase 
    words = letters_only.lower().split()
    
    # lemmatize
    lemmatizer = WordNetLemmatizer()
    tokens_lem = [lemmatizer.lemmatize(i) for i in words]
    
    # stop words
    meaningful_words = [w for w in words if not w in stops]
    
    # return as a string 
    return(" ".join(meaningful_words))

total_message = df.shape[0]
clean_message = []

print("Cleaning and parsing the message...")

j = 0
for message in df['text']:
    clean_message.append(message_to_words(message))
    
    # If the index is divisible by 100, print a message
    if (j+1) % 100 == 0:
        print(f'Comment {j+1} of {total_message}.')
    
    j += 1
    
    if j == total_message:
        print('Done.')

df = df.assign(cleaned_message = clean_message)
df.head(3)

df.shape

df = df.drop(columns=['Unnamed: 0', 'Unnamed: 0.1', 'username', 'text'])

df.shape

# Creating the Urgent vs. non-Urgent Vectors

**Helpful medium post**
https://medium.com/@belen.sanchez27/leveraging-social-media-to-map-disasters-74b4cc34848d

#Credit to NYC III group (Andrew Sternick, Nick Read, Preeya Sawadmanod), their code provided a better 
#visualization of the cosine similarity scores for each message vector so when I started to have issues 
#I found their code and applied it while I am tweeting my bags of words for a clearer understanding of the 
#cosine similarity scores 


#Function for vectorization of corpus 
def twitter_vector(vector):
    
    #Counter for number of words in corpus that exists in GoogleNews word list 
    count=0
   
    #Creating a template for cumulative corpus vector sum (300 dimensions)
    tweet_vector = np.zeros((1,300))
    
    #Iterating over each word in bag of word list 
    for word in vector:

        #Checking if word exists in GoogleNews word list
        if word in model.vocab:                    
            
            #Vectorizing the word if in list
            word_vector = model.word_vector(word)        
            
            #Updating counter
            count += 1
            
            #Updating cumulative vector sum 
            tweet_vector = tweet_vector + word_vector 

    #Computing average vector by taking cumulative vector sum and dividing it by number of words traced
    tweet_vector_score = tweet_vector/count
    
    #Using numpy to squeeze N-dimensional nested array object into a 1-D array 
    tweet_vector_score = np.squeeze(tweet_vector_score)
    
    return(tweet_vector_score)

# Applying function to vectorize both bag of words list
urgent_vec = twitter_vector(urgent)
nonurgent_vec = twitter_vector(non_urgent)

urgent_vec.shape

nonurgent_vec.shape


#Function for computing cosine similarity score 
def cos_sim_score(a,b): 
    
    #Calculating the dot product of two vectors
    dot = np.dot(a, b)
    
    #Calculating the magnitude of the vector 
    norma = np.linalg.norm(a)
    normb = np.linalg.norm(b)
    
    #Calculating cosine similarity
    cos = dot / (norma * normb)
    
    return cos

from sklearn.metrics.pairwise import cosine_similarity

#original copy - DO NOT CHANGE
#Empty lists
urgent_cos_sim = []
non_urgent_cos_sim = []

#Iterating through each tweet in out list of tokens
for tweet in df['cleaned_message']:
    
    #Call function 
    avg_vec_per_tweet = tweet_vector(tweet)
    
    #Creating new column in df for cosine simarlity score 
    urgent_cos_sim.append(cos_sim_score(avg_vec_per_tweet, urgent_vec))
    non_urgent_cos_sim.append(cos_sim_score(avg_vec_per_tweet, nonurgent_vec))

df['urgent_cos_sim'] = urgent_cos_sim
df['non_urgent_cos_sim'] = non_urgent_cos_sim
df.head()

#Creating Classification 
urgent = []

for rows in df.index:  
    if df['urgent_cos_sim'][rows] > df['non_urgent_cos_sim'][rows]:
        urgent.append(1)
    else: 
        urgent.append(0)

#Creating a new column
df['urgent'] = urgent

df['urgent'].value_counts(normalize=True)

#view urgent tweets
#df[df['urgent'] == 0].head(5)

#read in the original dataframe to merge back with the raw tweets
df_original = pd.read_csv("../datasets/raw_df.csv")

#df_original.head()

df_original = df_original[['text','username']]

df_original.head()

df = pd.merge(df, df_original, left_index=True, right_index=True)

df.sample(n = 5)

df_urgent = df.loc[df['urgent'] == 1]

df_nonurgent = df.loc[df['urgent'] == 0]

#export df to get lat long points 
df_urgent.to_csv('..//datasets/df_urgent.csv')

#export df to get lat long points 
df_nonurgent.to_csv('..//datasets/df_nonurgent.csv')
