==> code/05_create-coordinates.ipynb <==
import pandas as pd
import numpy as np
from random import random
import json
import math

# Mapping and Visuals
import geog
import geopandas as gpd
import shapely
from shapely.geometry import Point, Polygon, mapping

# POSSIBLE LOCATIONS TO USE FOR MAPPING, WITH COORDINATES
coords_dict = {
    "Paradise": (39.76, -121.62),
    "Campbellville": (40.02, -121.723),
    "Chico": (39.728, -121.837),
    "Junction House": (39.316, -120.81),
    "Dodgeland": (39.544, -121.906)
}

loc_dict = {
    "City": [
        "Paradise",
        "Cambellville",
        "Chico",
        "Junction House",
        "Dodgeland"
    ],
    "Coords": list(coords_dict.values())
}

df = pd.DataFrame(loc_dict)

df["long"] = [i[1] for i in df["Coords"]]
df["lat"] = [i[0] for i in df["Coords"]]
df.head()

# Reference to DSI NYC III Cohort

def rand_cluster(n,c,r): 
    # N = NUMBER COORDS TO GENERATE
    # C = "CENTER" AT WHICH EVERYTHING WILL CLUSTER
    # R = "RADIUS" 
    x,y = c 
    points = []
    for i in range(n):
        theta = 2 * math.pi * random()
        s = r * random()
        points.append((x+s*math.cos(theta), y+s*math.sin(theta)))
    return points
# https://stackoverflow.com/questions/44356063/how-to-generate-a-set-of-random-points-within-a-given-x-y-coordinates-in-an-x

### Non-Urgent Tweets

non_urgent = pd.read_csv("../projects/p5/iter1/Client-Project/datasets/df_nonurgent.csv") # READ IN
non_urgent.drop(columns=["Unnamed: 0"], inplace=True) # DROP UNNEEDED COLS

# CHECK OUTPUT
print(non_urgent.shape)
non_urgent.head()

# add coords to non_urgent
non_urgent_coords = rand_cluster(non_urgent.shape[0], df["Coords"][0], .22)
non_urgent_coords = pd.Series(non_urgent_coords)
non_urgent["coordinates"] = non_urgent_coords
non_urgent[["lat", "long"]] = pd.DataFrame(non_urgent["coordinates"].tolist(), index=non_urgent.index)
non_urgent.tail()

non_urgent = gpd.GeoDataFrame(non_urgent,
                              geometry=gpd.points_from_xy(non_urgent["long"], non_urgent["lat"])) 
non_urgent.head()

### Urgent Tweets

urgent = pd.read_csv("../projects/p5/iter1/Client-Project/datasets/df_urgent.csv") # READ IN
urgent.drop(columns=["Unnamed: 0"], inplace=True) # DROP UNNEEDED COLS

# CHECK OUTPUT
print(urgent.shape)
urgent.head()

# add coors urgent
urgent_coords = rand_cluster(urgent.shape[0], df["Coords"][0], .22)
urgent_coords = pd.Series(urgent_coords)
urgent["coordinates"] = urgent_coords
urgent[["lat", "long"]] = pd.DataFrame(urgent["coordinates"].tolist(), index=urgent.index)
urgent.tail()

urgent = gpd.GeoDataFrame(urgent,
                          geometry=gpd.points_from_xy(urgent["long"], urgent["lat"])) 
urgent.head()

### Save New DFs to File

# save the 2 new df's
non_urgent.to_csv("./datasets/non_urgent_with_coords.csv")
urgent.to_csv("./datasets/urgent_with_coords.csv")
