==> code/06_coodinates_prep.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"></ul></div>

import pandas as pd

df = pd.read_csv('./non_urgent_with_coords.csv')

df.isnull().sum()

df = df.rename(columns={"text": "tweet"})

df = df[['tweet', 'lat', 'long']]

#df = df.sample(n=1034)

df.head()

df.to_csv('./non_urgent_with_coords.csv')

df = pd.read_csv('./urgent_with_coords.csv')

df = df.rename(columns={"text": "tweet"})

df = df[['tweet', 'lat', 'long']]

df.to_csv('./urgent_with_coords.csv')
